refactor(arika): split element sets into no-alloc Sgp4Elements + ParsedElementSet
Rename the shared mean-element record `omm::Omm` to `elements::Sgp4Elements`
and move it, along with the format dispatcher (`detect`/`parse`/`Format`), out
of the `omm` module into a new `arika::elements` module. TLE (`tle`) and the
CCSDS OMM serializations (`omm::{json,kvn,xml}`) are now format decoders into
this shared type, so a TLE no longer parses into a type named after the OMM
format.

`Sgp4Elements` becomes a no-alloc `Copy` numeric mean-element set (catalog
number, epoch, the six SGP4 mean elements, B*). It lives below the `alloc`
feature gate so the SGP4 propagation path can be used in no_std builds without
an allocator. Owned satellite identity (`OBJECT_NAME`/`OBJECT_ID`) moves to the
alloc-gated `ParsedElementSet`, which the text parsers return (text decoding
needs `alloc` regardless).

Behavior-preserving rename + restructure: the moved detect/parse/keplerian
tests pin the existing parsing and conversion behavior. arika builds clean on
the no_std (no alloc), no_std+alloc, and std tiers.
