+++ arika/src/lib.rs
@@ -8,6 +8,7 @@ mod math;
 pub mod body;
 pub mod earth;
 pub mod eclipse;
+pub mod elements;
 pub mod epoch;
 pub mod frame;
 #[cfg(feature = "alloc")]
